+++ back/src/main/java/com/openclassrooms/starterjwt/services/UserService.java
@@ -1,7 +1,15 @@
 package com.openclassrooms.starterjwt.services;
 
+import com.openclassrooms.starterjwt.exception.BadRequestException;
+import com.openclassrooms.starterjwt.exception.NotAuthorizedException;
+import com.openclassrooms.starterjwt.exception.NotFoundException;
 import com.openclassrooms.starterjwt.models.User;
 import com.openclassrooms.starterjwt.repository.UserRepository;
+
+import java.util.Objects;
+
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -12,11 +20,36 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
+    public void create(User user) {
+        if (userRepository.existsByEmail(user.getEmail())) {
+            throw new BadRequestException("Error: Email is already taken!");
+        }
+        this.userRepository.save(user);
+    }
+
     public void delete(Long id) {
+        User user = this.findById(Long.valueOf(id));
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication()
+                .getPrincipal();
+        if (!Objects.equals(userDetails.getUsername(), user.getEmail())) {
+            throw new NotAuthorizedException();
+        }
         this.userRepository.deleteById(id);
     }
 
     public User findById(Long id) {
-        return this.userRepository.findById(id).orElse(null);
+        User user = this.userRepository.findById(id).orElse(null);
+        if (user == null) {
+            throw new NotFoundException();
+        }
+        return user;
+    }
+
+    public User findByEmail(String email) {
+        User user = this.userRepository.findByEmail(email).orElse(null);
+        if (user == null) {
+            throw new NotFoundException();
+        }
+        return user;
     }
 }
